### 2021-09-30 #119 @ThomasA
I can offer to review. I have taught Python for Scientific Computing at Aalborg University for several years.
