### 2021-10-02 #138 @whedon
[REVIEW]: Teaching Python for Data Science: Collaborative development of a modular & interactive curriculum
**Submitting author:** @bnwolford (<a href="http://orcid.org/0000-0003-3153-1552">Brooke N. Wolford</a>)
**Repository:** <a href="https://github.com/GWC-DCMB/curriculum-notebooks/" target ="_blank">https://github.com/GWC-DCMB/curriculum-notebooks/</a>
**Version:** v1.0.0

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".

- [x] #142

- [x] #154
