### 2021-10-08 GWC-DCMB/curriculum-notebooks#7 @lechten
JOSE: Software license does not contain copyright
This is related to your [JOSE submission](https://github.com/openjournals/jose-reviews/issues/138).

Your [license](https://github.com/GWC-DCMB/curriculum-notebooks/blob/main/LICENSE.md) mentions "The above copyright notice", which is not expressed clearly. I was expecting to see something like this: "Copyright <YEAR> <COPYRIGHT HOLDER>"
