### 2021-10-08 GWC-DCMB/curriculum-notebooks#8 @lechten
JOSE: Version information?
This is related to your [JOSE submission](https://github.com/openjournals/jose-reviews/issues/138).

One of my review checkboxes asks whether the your version matches v0.1.0. I see v0.1.1 as of June 18.
