### 2021-10-08 GWC-DCMB/curriculum-notebooks#9 @lechten
JOSE: Community guidelines?
This is related to your [JOSE submission](https://github.com/openjournals/jose-reviews/issues/138).

One of my review checkboxes asks for community guidelines. Do they exist? Some projects have a file CONTRIBUTING...
