### 2021-10-08 GWC-DCMB/curriculum-notebooks#10 @lechten
JOSE: Paper contents
This is related to your [JOSE submission](https://github.com/openjournals/jose-reviews/issues/138).

The paper is well written and addresses the questions relevant to JOSE. I've got some comments/questions:

At the end of the summary you write that anyone can contribute. I suggest to emphasize that you describe Open Educational Resources (OER) under free and open licenses. Actually, I wonder about the videos. They do not indicate license information, do they? Are they OER too? Are they only available on Youtube or also elsewhere for download and reuse?

Section "Collaborative Curriculum Development"
- You describe a process of quality control via PRs. I do not see those PRs in this repository. Are they publicly visible somewhere?
- JOSE papers prominently feature a link to the repository. Thus, the link at the end of the section may not be necessary. Besides, is this section really the best place for such a link?

I suggest to phrase learning objectives with action verbs (what will students be able to *do*?). The first objective is too vague in this regard, while the other two are fine.

You write about "paired programming". I suggest "pair programming".
