### 2021-10-08 815f578
Clarify MIT license copyright (resolves #7)
