### 2021-10-11 GWC-DCMB/curriculum-notebooks#17 @lechten
JOSE: Prerequisites and collaboration on notebooks?
This is related to your [JOSE submission](https://github.com/openjournals/jose-reviews/issues/138).

I did not use Google Colab before. After reading the paper I thought that the Python source code for the notebooks was available on GitHub (so that I could [retain/reuse/revise/remix/redistribute it as OER](https://opencontent.org/blog/archives/3221), without Google Colab). I may be missing something fundamental here, but this does not seem to be the case. E.g., how would I create a PR to address issue #14 and change `&` into `and`?

Besides, if I open a workbook via a link from the README, pressing "play" asks me to log in to Google. Is that a hard prerequisite to use the notebooks? If so, please document this.
