+++ README.md
@@ -34,37 +34,37 @@ Take a look at the [Lesson](Lessons/_Keys/pdf) & [Practice](Practices/_Keys/pdf)
 |   | Lesson Video | Lesson Notebook | Practice Notebook |
 |---|---|---|---|
 | **Module I: Jupyter Setup** |
-| 1 | https://youtu.be/plRRJ1zupgI | None, just watch the video. | [Practice01 Jupyter Setup](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Practices/Practice01_Jupyter-Setup.ipynb) |
+| 1 | https://youtu.be/plRRJ1zupgI | None, just watch the video. | [Practice01 Jupyter Setup](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/main/Practices/Practice01_Jupyter-Setup.ipynb) |
 | **Module II: Coding Fundamentals** |
-| 2 | https://youtu.be/czhcehpotos | [Lesson02 HelloWorld and Variables](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Lessons/Lesson02_HelloWorld_Variables.ipynb) | [Practice02 HelloWorld and Variables](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Practices/Practice02_HelloWorld_Variables.ipynb) |
-| 3 | https://youtu.be/cGVIyeGv2bw | [Lesson03 Variables and Types](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Lessons/Lesson03_Variables_Types.ipynb) | [Practice03 Variables and Types](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Practices/Practice03_Variables_Types.ipynb) |
-| 4 | https://youtu.be/xY51hfthvrw | [Lesson04 Lists Intro](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Lessons/Lesson04_Lists_Intro.ipynb) | [Practice04 Lists Intro](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Practices/Practice04_Lists_Intro.ipynb) |
-| 5 | https://youtu.be/g9U_q5yWjrQ | [Lesson05 Indexing](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Lessons/Lesson05_Indexing.ipynb) | [Practice05 Indexing](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Practices/Practice05_Indexing.ipynb) |
-| 6 |   | [Lesson06 2D Lists Intro](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Lessons/Lesson06_2D_Lists_Intro.ipynb)  (legacy lesson) | [Practice06 2D Lists Intro](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Practices/Practice06_2D_Lists_Intro.ipynb) |
-| 7 |   | [Lesson07 2D Lists Indexing](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Lessons/Lesson07_2D_Lists_Indexing.ipynb)  (legacy lesson) | [Practice07 2D Lists Indexing](https://colab.research.google.com/github/GWC-DCMB/curriculum-notebooks/blob/master/Practices/Practice07_2D_Lists_Indexing.ipynb) |
[…]
