### 2021-10-11 bedef7b
Change 'paired' to 'pair' programming
