### 2021-10-11 35e5a20
Revise learning objectives to have action verbs
