### 2021-10-12 GWC-DCMB/curriculum-notebooks#17 @lechten
Ah, I see, thank you for the clarification. I thought that lesson notebooks are complete, while practice notebooks have blanks. I'm closing this as the README explains that both types follow a "fill in the blanks" style.
