### 2021-11-04 #138 @labarba
hi @ThomasA – The JOSE paper will include the version in the metadata, as indicated at the top of this thread. Once any revisions to the submission are complete, we will ask the authors to mint a new tagged release, and update that info here running `@whedon set version as v#.#`, and everything will be neat and square then. (It's OK to leave that item unchecked until the end.)
