### 2021-11-18 GWC-DCMB/curriculum-notebooks#8 @lechten
Clarified [there](https://github.com/openjournals/jose-reviews/issues/138#issuecomment-961002510).
