### 2021-11-18 #138 @lechten
Among the issues that I opened, https://github.com/GWC-DCMB/curriculum-notebooks/issues/10 has received an MR whose comment reads "Partially addresses issue #10". Thus, I did not take a closer yet. @bnwolford What is your view of the situation?
