### 2021-11-18 fbb0245
State that our resources are OER
