### 2021-11-18 c158d24
Remove redundant GitHub link
