### 2021-11-23 GWC-DCMB/curriculum-notebooks#10 @kelly-sovacool
Thank you for your comments, Jens!

1. We added the OER phrase to the summary in PR #21.
1. CC-BY also applies to the videos. The descriptions on YouTube list the license (you have to click 'show more' to see it). We updated the language in our license file here to explicitly mention the videos in PR #21. Currently, they are only available on YouTube. Would you suggest making them available for download on another platform also?
1. We've gone through a couple rounds of reorganizing/restructuring since we began this project. In addition to in this repo, PRs can be viewed [here](https://github.com/GWC-DCMB/SummerExperience/pulls?q=is%3Apr+is%3Aclosed) and [here](https://github.com/GWC-DCMB/ClubCurriculum/pulls?page=1&q=is%3Apr+is%3Aclosed).
1. We removed the link to GitHub at the end of the Section "Collaborative Curriculum Development" since it's redundant and doesn't flow naturally.
1. We rephrased the learning objectives in PR #18.
1. We switched "paired programming" to "pair programming" in PR #18.
