### 2021-11-23 e44638f
Run build on pull requests
