### 2021-11-23 585819b
Run build on PRs but not pushes
