+++ .github/workflows/build.yml
@@ -9,19 +9,11 @@ on:
       - Lessons/_Keys/*.ipynb
       - Practices/*.ipynb
       - Practices/_Keys/*.ipynb
-  push:
-    branches:
-      - main
-    paths:
-      - Lessons/*.ipynb
-      - Lessons/_Keys/*.ipynb
-      - Practices/*.ipynb
-      - Practices/_Keys/*.ipynb
 
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: macos-latest
     strategy:
       max-parallel: 4
       matrix:
@@ -40,7 +32,6 @@ jobs:
         pip install jupyter nbconvert
     - name: Configure git
       run: |
-        git pull
         git config --local user.email "noreply@github.com"
         git config --local user.name "GitHub"
     - name: Clean colab metadata
