### 2021-11-23 #138 @zenalapp
We have now resolved all of the issues that have been opened by @lechten, including issue [#10](https://github.com/openjournals/jose-reviews/issues/10).
