### 2021-11-23 #138 @ThomasA
The "convert to issue" functionality in the review checklist does not seem to work as intended. Whedon auto-closes them immediately.
