### 2021-11-23 GWC-DCMB/curriculum-notebooks#27 @ThomasA
Seeking support
*This is in relation to the [JOSE review](https://github.com/openjournals/jose-reviews/issues/138).*

"Community guidelines" ask for guidelines on seeking support. While you address contributing and reporting issues excellently, I do not find seeking support explicitly mentioned. I would personally assume that I can try opening an issue asking for support on something, but my assumption may be wrong and in any case, it might be a good idea to also briefly mention in your README whether and how you can ask for support on the lesson material.
