### 2021-11-23 #138 @ThomasA
@labarba I just pressed the button, hoping it was a shortcut to open an issue in the reviewed repo, but I guessed wrong. I just don't understand the purpose of functionality for conveniently opening an issue that automatically gets closed immediately, but thst's probably just a standard GitHub feature of checklists :smile:
