### 2021-11-29 GWC-DCMB/curriculum-notebooks#10 @lechten
Thank you for the improvements and explanations!

Yes, I would indeed suggest to make the videos available elsewhere, but that is just my personal opinion and unrelated to the review process: I see Google as driver of [surveillance capitalism](https://en.wikipedia.org/wiki/Surveillance_capitalism) and I would not want to force students into that surveillance to use my OER.
