### 2021-12-02 GWC-DCMB/curriculum-notebooks#10 @kelly-sovacool
@lechten one of the options we were thinking of was to put the videos on Google Drive and make them viewable publicly because we have free space through our University. Although still via Google, I think that would still be an improvement over only having them on YouTube? We're interested if you have other ideas.
