+++ .github/workflows/build.yml
@@ -13,7 +13,7 @@ on:
 jobs:
   build:
 
-    runs-on: macos-latest
+    runs-on: ubuntu-latest
     strategy:
       max-parallel: 4
       matrix:
