### 2021-12-02 GWC-DCMB/curriculum-notebooks#10 @lechten
The [5 Rs of openness](https://opencontent.org/blog/archives/3221) require a download (retain) option, which is missing on YouTube. Yes, with Google Drive this requirement would be met.

I do not have own experiences in sharing videos, so the following are just vague suggestions. Maybe [OER Commons](https://www.oercommons.org/about) would be a good place. Also [Wikimedia Commons](https://en.wikipedia.org/wiki/Wikipedia:WikiProject_Screencast) and [Internet Archive](https://help.archive.org/hc/en-us/articles/360002360111-Uploading-A-Basic-Guide) host videos.
