### 2021-12-02 #138 @lechten
All issues raised by me have been resolved. Except for the "version" item, I just ticked off the remaining items on my list. I vote to accept this paper.
