### 2021-12-08 69415ed
Clarify how to seek support (resolves #27)
