+++ README.md
@@ -68,6 +68,11 @@ Take a look at the [Lesson](Lessons/_Keys/pdf) & [Practice](Practices/_Keys/pdf)
 
 ## Help & Contributing
 
+If you need help getting started or using the material, 
+feel free to open an [issue](https://github.com/GWC-DCMB/curriculum-notebooks/issues)
+or [send us an email](gwc.bioinfo-requests@umich.edu)
+and we'll be happy to help!
+
 If you come across a bug, open an [issue](https://github.com/GWC-DCMB/curriculum-notebooks/issues)
 and include a minimal reproducible example.
 
