### 2021-12-08 GWC-DCMB/curriculum-notebooks#27 @kelly-sovacool
We added an extra sentence to the README to clarify this: https://github.com/GWC-DCMB/curriculum-notebooks#help--contributing
