### 2021-12-09 GWC-DCMB/curriculum-notebooks#27 @ThomasA
Sounds good to me
