### 2021-12-10 a7a80ad
Stop compiling PDFs in build workflow
