### 2021-12-10 7b8d3a7
Update build workflow based on mikropml pr_build
https://github.com/SchlossLab/mikropml/blob/main/.github/workflows/pr_build.yml
