### 2021-12-13 #138 @JasonJWilliamsNY
@zenalapp your next step is to go ahead and archive on Zenodo, issue a tagged release on the repository and mint a DOI. Please me know if I can assist you in any way on this. Once you have that, I'll recommend accept.
