### 2021-12-13 #138 @kelly-sovacool
@JasonJWilliamsNY The new release is v1.0.0 with DOI: `10.5281/zenodo.5776898`
