### 2021-12-14 #138 @JasonJWilliamsNY
@kelly-sovacool I notice that the author info/order is different on the Zenodo archive vs. the paper (PDF). Can you update the Zenodo repository (or paper if it is incorrect)?
