### 2021-12-14 #138 @kelly-sovacool
@JasonJWilliamsNY thanks for catching that, I've fixed the author info in Zenodo. The paper is correct.
