### 2021-12-15 #138 @labarba
hi @bnwolford 👋  — we are ready to publish your JOSE contribution. Before doing so, I wonder if you are willing to give the paper one editing pass to reduce the word count and repetition. For example, I found that the paper explains _three times_ the live coding format, and the pivot to flipped learning after the pandemic. Surely this could be tightened a bit?
