### 2021-12-17 9619f92
Revert "Update paper.md"
This reverts commit d2eaf64622263625e7c2e0f3b02f69726593e811.

Brooke accidentally made these edits on the main branch.
