+++ paper/paper.md
@@ -134,7 +134,7 @@ negatively influence their math and science performance and deter them from
 pursuing STEM as a career [@hill_why_2010].
 The majority of our organization's founding graduate students (all women) began
 coding in our undergraduate careers or later.
-We wanted to provide a safe environment for local high school women to develop
+We wanted to provide a safe place for local high school women to develop
 confidence in themselves and their computational skills before college, and be
 exposed to successful women role models in STEM to counter negative stereotypes.
 
@@ -167,9 +167,9 @@ Data Scientist was rated the #1 job in America by Glassdoor in 2016-2019, #3 in
 Furthermore, Python is the most popular programming language according to the
 PYPL PopularitY of Programming Language Index [@noauthor_pypl_nodate].
 Therefore, we believe career exploration in data science using the Python
-programming language will optimally prepare our learners for careers that
+programming language will optimally prepare our students for careers that
 provide financial stability and upward economic mobility.
-By leveraging the data science expertise of our Club facilitators (hereafter termed instructors), we created a
+By leveraging the data science expertise of our Club facilitators, we created a
 specialized curriculum focused on computational data science in the Python
 programming language.
 
@@ -177,10 +177,11 @@ Girls Who Code encourages participants to learn programming skills while working
 on an Impact Project website or application throughout the Club
 [@hq_gallery_2021].
 We created an open source Data Science curriculum that teaches the requisite
-Python and statistics skills to complete a Capstone Project, where learners
+Python and statistics skills to complete a Capstone Project, where students
 explore, analyze, and present a data set of their choosing.
-Using this curriculum, we employ participatory live coding, where learners type and run code along with the
-instructor in real time.
+Using this curriculum, we employ participatory live coding as used by The
+Carpentries, which is an effective method that engages learners
+[@wilson_software_2016; @nederbragt_ten_2020].
 Using paired activities, our curriculum follows the "I do, we do, you do"
 didactic paradigm [@fisher_better_2013].
 We provide open source resources for both in-person and virtual versions of our
@@ -192,8 +193,8 @@ coding for data science.
 # Collaborative Curriculum Development
 
 We assembled a team of volunteers involved in our club to develop a custom
-curriculum to teach introductory Python for data science.
-We chose the content based on what our learners would need to learn
+curriculum to teach introductory Python for data science via live coding.
+We chose the content based on what our students would need to learn
 to complete a small data analysis project and communicate their findings
 to their peers.
 We divided the content by topic into Jupyter notebooks for each lesson, with
@@ -220,7 +221,14 @@ Following the onset of the COVID-19 pandemic, we quickly pivoted our club to a
 virtual format.
 In preparation for the 2020 Summer Experience, we switched to a flipped
 classroom style following feedback from our club participants that it was too
-difficult to follow along live coding via Zoom (see (##Instructional Design)).
+difficult to follow along live coding via Zoom.
+We recorded facilitators teaching the lesson notebooks as if they were live
+coding, then shared them along with a link to the lesson notebooks for students
+to code along with while watching the videos.
+Each video shows the Jupyter notebook alongside the facilitator themselves
+teaching.
+This format allowed students to learn at their own pace, then ask questions and
+practice when we met together virtually.
 
 # Curriculum
 
@@ -243,7 +251,7 @@ The learning objectives of this curriculum are:
 3. Apply data analysis to real world data to answer scientific questions.
 4. Create informative summary statistics and data visualizations in Python.
 
-These skills provide a solid foundation for basic data analysis in Python. Participation in our program exposes learners to the many ways coding and data science can be impactful across many disciplines.
[…]
