+++ paper/paper.md
@@ -220,7 +220,7 @@ Following the onset of the COVID-19 pandemic, we quickly pivoted our club to a
 virtual format.
 In preparation for the 2020 Summer Experience, we switched to a flipped
 classroom style following feedback from our club participants that it was too
-difficult to follow along live coding via Zoom (see (##Instructional Design)).
+difficult to follow along live coding via Zoom (see [Instructional Design](#instructional-design)).
 
 # Curriculum
 
@@ -228,11 +228,6 @@ Our curriculum was designed for high school students with no prior coding
 experience who are interested in learning Python programming for data science.
 However, this course material would be useful for anyone interested in
 teaching or learning basic programming for data analysis.
-
-Our curriculum features short lessons to deliver course material in “bite sized”
-chunks, followed by practices to solidify the learners' understanding.
-Pre-recorded videos of lessons enable effective virtual learning and flipped
-classroom approaches.
 
 ## Learning Objectives
 
@@ -246,7 +241,6 @@ The learning objectives of this curriculum are:
 These skills provide a solid foundation for basic data analysis in Python. Participation in our program exposes learners to the many ways coding and data science can be impactful across many disciplines.
 
 ## Course Content
-
 
 Our curriculum design consists of 27 lessons broken up into 5 modules that cover
 Jupyter notebook setup, Python coding fundamentals, use of essential data
@@ -267,7 +261,7 @@ working on a data science project using a real world
 dataset of their choosing.
 While more time is dedicated to lessons early in the program, the formal
 curriculum tapers off until the learners are solely applying their skills to the
-data science project.  
+data science project.
 Through this Capstone Project, learners gain practical experience with each
 skill as they learn it in the lessons; including importing and cleaning data,
 data visualization, and basic statistical analyses.
@@ -286,18 +280,12 @@ the material and deal with errors and bugs as they arise.
 practice, providing further opportunity for practical experience implementing
 the coding skill at hand and testing learners' understanding of the content.
 
-This curriculum was originally developed for in-person instruction, but the
-onset of the COVID-19 pandemic necessitated restructuring to a virtual format.
-To better facilitate virtual instruction, we switched to a flipped classroom.
-
-We provide video recordings of instructors teaching the lesson notebooks as if they were live
-coding in the classroom.
-Learners then watch videos of instructors explaining the material through "live"
+To better facilitate virtual instruction during the COVID-19 pandemic, we switched to a flipped classroom.
+Prior to meeting, learners watch videos of instructors explaining the material through "live"
 coding and code along in the lesson notebook with while watching the video.
 Each video shows the Jupyter notebook alongside the instructor themselves
 teaching. Learners then complete a practice notebook corresponding to the lesson.
-This format allowed learners to explore the material individually before meeting together.
-Instructors then spend the virtual meeting time answering questions and reviewing the
+During the virtual meeting time, instructors answer questions and review the
 core concepts in the practice exercises. This virtual format is especially beneficial because it 1) allows learners to
 learn at their own pace, and 2) enables dissemination of our curriculum to a
 wider audience interested in learning introductory Python programming for data
@@ -327,7 +315,7 @@ them to successful women in STEM.
 We have used this curriculum to teach the Data Science Summer Experience and
 Girls Who Code Club in person in 2019 and virtually in 2020-2021.
 For the in-person instances, we taught the curriculum through participatory live
-coding, and for virtual instances, we used a flipped classroom approach (see (##Instructional Design)).
[…]
