### 2021-12-17 c5eb651
Move fig 2 down
