### 2021-12-17 1ec58ac
Remove duplicated paragraph
