### 2021-12-17 d796f97
Streamline a bit more
