+++ paper/paper.md
@@ -125,10 +125,9 @@ percentage of females applying directly to the Bioinformatics PhD program was
 first or second choice in the Program in Biomedical Sciences, U-M's biomedical
 PhD umbrella program was 41%.
 
-Previous research on women's educational experiences in science, technology, engineering,
-and mathematics (STEM) have produced
-various explanations for persistent gender disparities
-[@benbow_gender_2016].
+Previous research on women's educational experiences in science, technology,
+engineering, and mathematics (STEM) have produced various explanations for
+persistent gender disparities [@benbow_gender_2016].
 One explanation is that women often experience stereotype threats that
 negatively influence their math and science performance and deter them from
 pursuing STEM as a career [@hill_why_2010].
@@ -152,9 +151,9 @@ Participants in the Club attend weekly meetings at the University of Michigan
 (when the club is run in person rather than virtually), and are thus largely
 high school women from the Ann Arbor area.
 In 2019 we launched our own summer program, the Data Science Summer Experience.
-When held in person, the Summer Experience
-is hosted in Detroit to provide the opportunity
-for high school women outside of Ann Arbor to learn coding skills in an inclusive environment.
+When held in person, the Summer Experience is hosted in Detroit to provide the
+opportunity for high school women outside of Ann Arbor to learn coding skills in
+an inclusive environment.
 
 The national Girls Who Code organization provides a curriculum that teaches
 website and application development through programming languages like HTML and
@@ -169,9 +168,9 @@ PYPL PopularitY of Programming Language Index [@noauthor_pypl_nodate].
 Therefore, we believe career exploration in data science using the Python
 programming language will optimally prepare our learners for careers that
 provide financial stability and upward economic mobility.
-By leveraging the data science expertise of our Club facilitators (hereafter termed instructors), we created a
-specialized curriculum focused on computational data science in the Python
-programming language.
+By leveraging the data science expertise of our Club facilitators (hereafter
+termed instructors), we created a specialized curriculum focused on
+computational data science in the Python programming language.
 
 Girls Who Code encourages participants to learn programming skills while working
 on an Impact Project website or application throughout the Club
@@ -179,8 +178,8 @@ on an Impact Project website or application throughout the Club
 We created an open source Data Science curriculum that teaches the requisite
 Python and statistics skills to complete a Capstone Project, where learners
 explore, analyze, and present a data set of their choosing.
-Using this curriculum, we employ participatory live coding, where learners type and run code along with the
-instructor in real time.
+Using this curriculum, we employ participatory live coding, where learners type
+and run code along with the instructor in real time.
 Using paired activities, our curriculum follows the "I do, we do, you do"
 didactic paradigm [@fisher_better_2013].
 We provide open source resources for both in-person and virtual versions of our
@@ -220,7 +219,8 @@ Following the onset of the COVID-19 pandemic, we quickly pivoted our club to a
 virtual format.
 In preparation for the 2020 Summer Experience, we switched to a flipped
 classroom style following feedback from our club participants that it was too
-difficult to follow along live coding via Zoom (see [Instructional Design](#instructional-design)).
+difficult to follow along live coding via Zoom 
+(see [Instructional Design](#instructional-design)).
 
 # Curriculum
 
@@ -234,11 +234,14 @@ teaching or learning basic programming for data analysis.
 The learning objectives of this curriculum are:
 
 1. Write code in Python with correct syntax and following best practices.
-2. Implement fundamental programming concepts when presented with a programmatic problem set.
+2. Implement fundamental programming concepts when presented with a programmatic
+   problem set.
 3. Apply data analysis to real world data to answer scientific questions.
[…]
