### 2021-12-17 #138 @zenalapp
Thanks for catching that @labarba, we have streamlined the paper and pushed those changes.
