### 2021-12-17 #138 @labarba
Congratulations, folks! Your JOSE article is published! 🚀

Huge thanks to our Editor: @JasonJWilliamsNY and Reviewers: @lechten, @ThomasA — we couldn't do this without you 🙏
